requirements.txt
pyrogram
tgcrypto
nest_asyncio
apscheduler
streamlit_javascript
pymongo
git+https://github.com/rthalley/dnspython
uvloop